+++ src/components/RadioGroup/RadioGroup.stories.tsx
@@ -19,7 +19,7 @@
 import { Meta, StoryObj } from '@storybook/react'
 import { action } from '@storybook/addon-actions'
 
-import { RadioGroupOption, RadioGroupProps } from './RadioGroup.props'
+import { RadioGroupOption, RadioGroupProps } from './props'
 import { RadioGroup } from './RadioGroup'
 
 const baseOptions: RadioGroupOption[] = [
